extract tile
