+++ src/tile.hpp
@@ -1,0 +1,9 @@
+#ifndef WFC_TILE_HPP
+#define WFC_TILE_HPP
+
+#include "constants.hpp"
+
+Tile rotate(const Tile& in_tile, const size_t tile_size);
+
+#endif
+
